config/settings/base.py
import environ

# Three levels up from pathways-backend/config/settings/base.py gives pathways-backend/
ROOT_DIR = environ.Path(__file__) - 3
APPS_DIR = ROOT_DIR.path('main')

env = environ.Env()
READ_DOT_ENV_FILE = env.bool('DJANGO_READ_DOT_ENV_FILE', default=True)

if READ_DOT_ENV_FILE:
    # Operating System Environment variables have precedence over variables defined in
    # the .env file, that is to say variables from the .env files will only be used if
    # not defined as environment variables.
    env_file = str(ROOT_DIR.path('.env'))
    env.read_env(env_file)

DJANGO_APPS = [
    'django.contrib.auth',
    'django.contrib.contenttypes',
    'django.contrib.sessions',
    'django.contrib.sites',
    'django.contrib.messages',
    'django.contrib.staticfiles',
    'django.contrib.admin',
]

THIRD_PARTY_APPS = [
    'crispy_forms',  # Form layouts
    'allauth',  # registration
    'allauth.account',  # registration
    'allauth.socialaccount',  # registration
    'corsheaders',
    'rest_framework',
    'drf_yasg',
    'parler',
    'phonenumber_field',
]

LOCAL_APPS = [
    'common.apps.CommonConfig',
    'bc211.apps.Bc211Config',
    'newcomers_guide.apps.NewcomersGuideConfig',
    'human_services.locations.apps.LocationsConfig',
    'human_services.organizations.apps.OrganizationsConfig',
    'human_services.services.apps.ServicesConfig',
    'human_services.addresses.apps.AddressesConfig',
    'human_services.phone_numbers.apps.PhoneNumbersConfig',
    'users.apps.UsersConfig',
    'taxonomies.apps.TaxonomiesConfig',
    'translation.apps.ContentTranslationToolsConfig',
]

INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS

REST_FRAMEWORK = {
    'DEFAULT_PAGINATION_CLASS': 'common.view.Pagination',
    'PAGE_SIZE': 30,
    'ORDERING_PARAM': 'sort_by',
    'DEFAULT_FILTER_BACKENDS': ['common.filters.SearchFilter', 'common.filters.MultiFieldOrderingFilter', ],
}

MIDDLEWARE = [
    'django.middleware.security.SecurityMiddleware',
    'django.contrib.sessions.middleware.SessionMiddleware',
    'corsheaders.middleware.CorsMiddleware',
    'django.middleware.locale.LocaleMiddleware',
    'django.middleware.common.CommonMiddleware',
    'django.middleware.csrf.CsrfViewMiddleware',
    'django.contrib.auth.middleware.AuthenticationMiddleware',
    'django.contrib.messages.middleware.MessageMiddleware',
    'django.middleware.clickjacking.XFrameOptionsMiddleware',
]

MIGRATION_MODULES = {
    'sites': 'main.contrib.sites.migrations'
}

DEBUG = env.bool('DJANGO_DEBUG', False)
CORS_ORIGIN_ALLOW_ALL = True

FIXTURE_DIRS = (
    str(APPS_DIR.path('fixtures')),
)

EMAIL_BACKEND = env('DJANGO_EMAIL_BACKEND', default='django.core.mail.backends.smtp.EmailBackend')

ADMINS = [
    ("""PeaceGeeks""", 'rasmus@peacegeeks.org'),
]

MANAGERS = ADMINS
TIME_ZONE = 'UTC'
LANGUAGE_CODE = 'en-us'
SITE_ID = 1
USE_I18N = True
USE_L10N = True
USE_TZ = True

TEMPLATES = [
    {
        'BACKEND': 'django.template.backends.django.DjangoTemplates',
        'DIRS': [
            str(APPS_DIR.path('templates')),
        ],
        'OPTIONS': {
            'debug': DEBUG,
            'loaders': [
                'django.template.loaders.filesystem.Loader',
                'django.template.loaders.app_directories.Loader',
            ],
            'context_processors': [
                'django.template.context_processors.debug',
                'django.template.context_processors.request',
                'django.contrib.auth.context_processors.auth',
                'django.template.context_processors.i18n',
                'django.template.context_processors.media',
                'django.template.context_processors.static',
                'django.template.context_processors.tz',
                'django.contrib.messages.context_processors.messages',
            ],
        },
    },
]

STATIC_ROOT = str(ROOT_DIR('staticfiles'))
STATIC_URL = '/static/'
STATICFILES_DIRS = [
    str(APPS_DIR.path('static')),
]

STATICFILES_FINDERS = [
    'django.contrib.staticfiles.finders.FileSystemFinder',
    'django.contrib.staticfiles.finders.AppDirectoriesFinder',
]

MEDIA_ROOT = str(APPS_DIR('media'))
MEDIA_URL = '/media/'
ROOT_URLCONF = 'config.urls'

VERBOSE_LOGGING_WITH_INFO = {
    'handlers': ['verbose-console'],
    'level': 'INFO',
}

LOGGING = {
    'version': 1,
    'disable_existing_loggers': False,
    'formatters': {
        'verbose-format': {
            'format': '%(levelname)s %(module)s %(message)s'
        },
    },
    'handlers': {
        'console': {
            'class': 'logging.StreamHandler',
        },
        'verbose-console': {
            'class': 'logging.StreamHandler',
            'formatter': 'verbose-format'
        },
    },
    'loggers': {
        'django': {
            'handlers': ['console'],
            'level': 'INFO',
        },
        'bc211': VERBOSE_LOGGING_WITH_INFO,
        'common': VERBOSE_LOGGING_WITH_INFO,
        'translation': VERBOSE_LOGGING_WITH_INFO,
        'human_services': VERBOSE_LOGGING_WITH_INFO,
        'users': VERBOSE_LOGGING_WITH_INFO,
    },
}

PARLER_DEFAULT_LANGUAGE_CODE = 'en'
PARLER_LANGUAGES = {
    1: (
        {'code': 'en', },
        {'code': 'fr', },
        {'code': 'nl', },
    ),
    'default': {
        'fallbacks': ['en'],
        'hide_untranslated': False,
    }
}

PARLER_PO_CONTACT = 'translations@peacegeeks.org'

DATABASES = {
    'default': {
        'ENGINE': 'django.contrib.gis.db.backends.postgis',
        'NAME': env('POSTGRES_DATABASE', default='pathways_local'),
        'USER': env('POSTGRES_USER', default='pathways'),
        'PASSWORD': env('POSTGRES_PASSWORD', default=''),
        'ATOMIC_REQUESTS': True,
    }
}

GDAL_LIBRARY_PATH = env('GDAL_LIBRARY_PATH', default='')
GEOS_LIBRARY_PATH = env('GEOS_LIBRARY_PATH', default='')

WSGI_APPLICATION = 'config.wsgi.application'

PASSWORD_HASHERS = [
    'django.contrib.auth.hashers.Argon2PasswordHasher',
    'django.contrib.auth.hashers.PBKDF2PasswordHasher',
    'django.contrib.auth.hashers.PBKDF2SHA1PasswordHasher',
    'django.contrib.auth.hashers.BCryptSHA256PasswordHasher',
    'django.contrib.auth.hashers.BCryptPasswordHasher',
]

AUTH_PASSWORD_VALIDATORS = [
    {
        'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator',
    },
    {
        'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator',
    },
    {
        'NAME': 'django.contrib.auth.password_validation.CommonPasswordValidator',
    },
    {
        'NAME': 'django.contrib.auth.password_validation.NumericPasswordValidator',
    },
]

AUTHENTICATION_BACKENDS = [
    'django.contrib.auth.backends.ModelBackend',
    'allauth.account.auth_backends.AuthenticationBackend',
]

ACCOUNT_AUTHENTICATION_METHOD = 'username'
ACCOUNT_EMAIL_REQUIRED = True
ACCOUNT_EMAIL_VERIFICATION = 'mandatory'

ACCOUNT_ALLOW_REGISTRATION = env.bool('DJANGO_ACCOUNT_ALLOW_REGISTRATION', True)
ACCOUNT_ADAPTER = 'users.adapters.AccountAdapter'
SOCIALACCOUNT_ADAPTER = 'users.adapters.SocialAccountAdapter'

AUTH_USER_MODEL = 'users.User'
LOGIN_REDIRECT_URL = 'users:redirect'
LOGIN_URL = 'account_login'

AUTOSLUG_SLUGIFY_FUNCTION = 'slugify.slugify'

ADMIN_URL = r'^admin/'

SRID = 4326
